pyproject.toml
[project]
name = "agentship"
version = "0.1.0"
description = "AgentShip - An Agent Shipping Kit. Production-ready framework for building, deploying, and operating AI agents."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core AI Agent Platform Dependencies
    "google-adk>=1.14.1",
    "litellm>=1.77.1",
    "fastapi>=0.116.2",
    "uvicorn>=0.36.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    
    # HTTP and Networking
    "aiohttp>=3.12.15",
    "httpx>=0.28.1",
    "requests>=2.32.3",
    
    # Configuration and Environment
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    
    # CLI and Utilities
    "click>=8.1.7",
    "typing-extensions>=4.12.2",
    
    # Google Cloud Dependencies
    "google-api-python-client>=2.150.0",
    "google-cloud-aiplatform>=1.115.0",
    "google-cloud-bigtable>=2.20.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-spanner>=3.20.0",
    "google-cloud-storage>=2.18.0",
    "google-genai>=0.8.0",
    
    # Authentication and Security
    "google-auth>=2.35.0",
    "authlib>=1.3.2",
    
    # Database and Storage
    "sqlalchemy>=2.0.36",
    "sqlalchemy-spanner>=1.0.0",
    "psycopg2-binary>=2.9.9",
    
    # Observability and Monitoring
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-gcp-trace>=1.7.0",
    
    # AI and ML
    "openai>=1.108.1",
    "tiktoken>=0.8.0",
    "tokenizers>=0.21.0",
    
    # Web Framework Dependencies
    "starlette>=0.41.6",
    "sse-starlette>=2.1.0",
    "websockets>=14.1",
    
    # Utilities
    "anyio>=4.6.2",
    "tenacity>=9.0.0",
    "watchdog>=6.0.0",
    "tzlocal>=6.0",
    "python-dateutil>=2.9.0",
    "jinja2>=3.1.4",
    "jsonschema>=4.23.0",
    "packaging>=24.2",
    "protobuf>=5.28.3",
    "proto-plus>=1.23.4",
    "shapely>=2.0.6",
    "graphviz>=0.20.3",
    "langgraph",
    "langgraph-checkpoint-postgres",
    "langchain-core",
    
    # MCP (Model Control Plane)
    "mcp>=1.14.1",
    
    # Production Server
    "gunicorn>=23.0.0",
]
